streamlit_app/requirements.txt
streamlit
databricks-sql-connector